Add paginationOptions to Ink.Table
When a table is initialized via JS (new Ink.UI.Table), the navigation, in the case where we want
pagination is created and initialized for us. But we may want to change something in it so this patch
allows to pass the pagination options to the Table so that when the pagination UI is created it can
have the options we want.