Use existing <nav> if not in markup mode
Currently, when in markup mode, when creating a table, it reuses the <nav> element right after the table.
This allows us to customize it (using classes) to our needs. If not using markup mode, it is never tested
for a <nav> element after the table so this patch just mimics the markup-style initialization.