feat: change handleSelector defaults
Minimizes an issue when your elements have anchors or clickable elements inside them.

Quirk: When a mouseup/touchend happens, the element below your cursor is a *copy* of the element you triggered the mousedown/touchstart on. This means it's NOT the same element and the click event will NOT trigger as expected.

When an handleSelector is defined, the copy/moving logic will only start if the element you mousedown/touchstart matches it. 

You can now click on anchors and buttons inside sortable items as they won't, by default, trigger the moving logic of that item.